+++ src/PEAKLib.Items/ItemContent.cs
@@ -19,9 +19,15 @@ public class ItemContent(Item item) : IContent<ItemContent>, IItemContent
     public RegisteredContent<ItemContent> Register(ModDefinition owner)
     {
         var registered = ContentRegistry.Register(this, owner);
+
+        var modItemComponent = item.GetComponent<ModItemComponent>();
+        if (modItemComponent != null)
+            modItemComponent.InitializeModItem(owner);
+
 #if !UNITY_EDITOR
         NetworkPrefabManager.RegisterNetworkPrefab(owner, "0_Items/", item.gameObject);
         s_RegisteredItems.Add(registered);
+        ItemRegistrar.RegisterIfTooLate(registered);
 #endif
         return registered;
     }