+++ src/PEAKLib.Items/ModItemData.cs
@@ -1,0 +1,60 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Zorro.Core.Serizalization;
+
+namespace PEAKLib.Items;
+
+/// <summary>
+/// Similar to other network-serialized types like <see cref="global::IntItemData"/>, but stores mod-indexed bytes as entries.
+/// </summary>
+internal class ModItemData : DataEntryValue
+{
+    /// <summary>
+    /// Value for input to serialization and output from deserialization
+    /// </summary>
+    public Dictionary<int, byte[]> Value = [];
+
+    /// <inheritdoc/>
+    public override void SerializeValue(BinarySerializer serializer)
+    {
+        // number of entries
+        serializer.WriteInt(Value.Count);
+        foreach (var pair in Value)
+        {
+            // mod ID
+            serializer.WriteInt(pair.Key);
+            // bytes length
+            serializer.WriteInt(pair.Value.Length);
+            // bytes
+            serializer.WriteBytes(pair.Value);
+        }
+    }
+
+    /// <inheritdoc/>
+    public override void DeserializeValue(BinaryDeserializer deserializer)
+    {
+        int count = deserializer.ReadInt();
+        for (int i = 0; i < count; i++)
+        {
+            int mod = deserializer.ReadInt();
+            int len = deserializer.ReadInt();
+            Value[mod] = deserializer.ReadBytes(len);
+        }
+    }
+
+    /// <inheritdoc/>
+    public override string ToString()
+    {
+        StringBuilder sb = new();
+        // output as hexadecimal
+        sb.Append($"ModItemData({Value.Count}) {{ ");
+        foreach (var pair in Value)
+        {
+            string hex = BitConverter.ToString(pair.Value);
+            sb.Append($"{pair.Key}=[{hex}], ");
+        }
+        sb.Append("}");
+        return sb.ToString();
+    }
+}