+++ src/PEAKLib.Core/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-07-14
+
+### Added
+
+- ModDefinition now overrides `GetHashCode` to return the hash code of the Id string.
+
+## [1.1.2] - 2025-07-12
+
+### Fixed
+
+- Invalid Content or bad Content implementations should no longer break the BundleLoader.
+
 ## [1.1.1] - 2025-07-09
 
 ### Fixed