+++ src/PEAKLib.Items/CHANGELOG.md
@@ -5,7 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2025-08-07
+## [1.3.0] - 2025-07-14
+
+### Added
+
+- Items can now be registered later than before.
+- More shaders will be swapped to the correct ones from the game than before.
+
+## [1.2.0] - 2025-07-14
+
+### Added
+
+- Custom item data is now supported for items.
+
+## [1.1.2] - 2025-07-12
+
+### Fixed
+
+- Invalid Content no longer breaks the BundleLoader.
+
+## [1.0.0] - 2025-07-08
 
 ### Added